Router fix 0.2
