+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './styles/app/App.css';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import { BrowserRouter as Router, Switch, Route, Redirect } from 'react-router-dom';
 import WorkflowRouter from './pages/workflow/WorkflowRouter';
 import AuthRouter from './pages/auth/AuthRouter';
 import Error404 from './pages/Error/404page';
@@ -11,7 +11,8 @@ function App() {
       <Switch>
         <Route path='/workflow' component={WorkflowRouter} />
         <Route path="/auth" component={AuthRouter}/>
-        <Route path="*" component={Error404}/>
+        <Route path="/not-found" component={Error404}/>
+        <Redirect to="/not-found"/>
       </Switch>
     </Router>
   );
