+++ src/pages/workflow/Dashboard.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Grid from "@material-ui/core/Grid";
 import Paper from "@material-ui/core/Paper";
-import WorkflowStyle from "./WorkflowStyle";
+import WorkflowStyle from "../../components/workflow/WorkflowStyle";
 import clsx from "clsx";
 
 export default function Dashboard (){
