+++ src/pages/workflow/WorkflowRouter.js
@@ -1,24 +1,28 @@
 import React from 'react'
-import SideMenu from './SideMenu'
+import SideMenu from '../../components/workflow/SideMenu'
 import Dashboard from './Dashboard'
 import AdminUsers from './AdminUsers'
 import Defects from './Defects'
 import WorkerUsers from './WorkerUsers'
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
+// import Error404 from '../Error/404page'
+import { Route, Switch, Redirect } from 'react-router-dom';
+
 
 export default function WorkflowRouter() {
     return (
-        <Router>
             <SideMenu
                 body={
                     <Switch>
-                        <Route path="/workflow/dashboard" component={Dashboard} />
-                        <Route path="/workflow/admin-users" component={AdminUsers}/>
-                        <Route path="/workflow/defects" component={Defects} />
-                        <Route path="/workflow/worker-users" component={WorkerUsers} />
+                        <Route exact path="/workflow/dashboard" component={Dashboard} />
+                        <Route exact path="/workflow/admin-users" component={AdminUsers} />
+                        <Route exact path="/workflow/defects" component={Defects} />
+                        <Route exact path="/workflow/worker-users" component={WorkerUsers} />
+                        <Route exact path="/workflow">
+                            <Redirect to="/workflow/dashboard" />
+                        </Route>
+                        <Redirect to="/"/>
                     </Switch>
                 }
             />
-        </Router>
     )
 }
